v1.0.9.277 - remove import command. goodbye, v0.8

+++ src/net/slipcor/pvparena/arena/Arena.java
@@ -1348,7 +1348,7 @@ public class Arena {
 								.setLastDamageCause(
 										new EntityDamageEvent(players.get(
 												playerLoc).get(),
-												DamageCause.CUSTOM, 1000));
+												DamageCause.CUSTOM, (double) 1000));
 						players.get(playerLoc)
 								.get()
 								.damage(getArenaConfig().getInt(
